Moves invoice service to shared lib directory
