+++ test/modules/billing/services/batch-service.js
@@ -32,7 +32,7 @@ const batchService = require('../../../../src/modules/billing/services/batch-ser
 const billingVolumesService = require('../../../../src/modules/billing/services/billing-volumes-service');
 const invoiceAccountsService = require('../../../../src/lib/services/invoice-accounts-service');
 const invoiceLicencesService = require('../../../../src/modules/billing/services/invoice-licences-service');
-const invoiceService = require('../../../../src/modules/billing/services/invoice-service');
+const invoiceService = require('../../../../src/lib/services/invoice-service');
 const licencesService = require('../../../../src/lib/services/licences');
 const transactionsService = require('../../../../src/modules/billing/services/transactions-service');
 const { createBatch } = require('../test-data/test-billing-data');
